pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-server-architect"
version = "0.1.0"
description = "AI Software Architect MCP Server that generates PRDs based on codebase analysis"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "ArchitectAI", email = "example@example.com"},
]
dependencies = [
    "mcp-sdk[cli]>=0.7.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
keywords = ["mcp", "llm", "architecture", "prd", "design"]
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
mcp-server-architect = "mcp_server_architect.__main__:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
